confirme/sujet/requirements.txt
customtkinter>=5.0.2
matplotlib>=3.5.3
tkinter_tooltip>=2.1.0
Pillow>=9.5.0
